# openapi.yaml
openapi: 3.0.3
info:
  title: SmartDrop API
  description: |
    SmartDrop backend services — price oracle, webhook delivery, health monitoring, and indexing.

    This specification covers all current REST endpoints and planned future endpoints.
    The public HTTP API is versioned by URL path under `/api/v1`; breaking
    changes are introduced in a new versioned path rather than in place.
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: SmartDrop Labs
    url: https://github.com/SmartDropLabs
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.smartdrop.app
    description: Production (planned)
security: []
paths:
...
  /api/v1/keys:
    get:
      operationId: listApiKeys
      summary: List API keys
      description: Returns sanitized API key metadata. Secret key material and key hashes are never returned.
      tags:
        - API Keys
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Existing API keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyListResponse"
              example:
                keys:
                  - id: key_1234567890abcdef
                    label: Production webhook worker
                    key_prefix: 3b28a419
                    scopes:
                      - webhooks:write
                    created_at: "2026-06-27T12:00:00.000Z"
                    last_used_at: null
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createApiKey
      summary: Create API key
      description: Creates a new API key and returns the plaintext secret once. Store it securely; subsequent responses only include metadata.
      tags:
        - API Keys
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyCreateRequest"
            example:
              label: Production webhook worker
              scopes:
                - webhooks:write
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreateResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/keys/{id}:
    delete:
      operationId: revokeApiKey
      summary: Revoke API key
      description: Revokes an API key by ID and returns the sanitized record that was removed.
      tags:
        - API Keys
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: ^key_
          description: API key ID
          example: key_1234567890abcdef
      responses:
        "200":
          description: API key revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyRevokeResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/airdrops/{id}/recipients:
    post:
      operationId: addAirdropRecipients
      summary: Add recipients to an airdrop
      description: |
        Adds recipients to an airdrop via JSON body or CSV file upload.
        Duplicate addresses within the request are rejected.
      tags:
        - Airdrops
      parameters:
        - $ref: "#/components/parameters/AirdropIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRecipientsRequest"
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file with address and amount columns
      responses:
        "201":
          description: Recipients added
          content:
            application/json:
              schema:
                type: object
                properties:
                  added:
                    type: integer
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Duplicate recipient addresses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listAirdropRecipients
      summary: List airdrop recipients
      description: Returns a paginated list of recipients for an airdrop.
      tags:
        - Airdrops
      parameters:
        - $ref: "#/components/parameters/AirdropIdParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Paginated list of recipients
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Recipient"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/airdrops/{id}/status:
    get:
      operationId: getAirdropOnchainStatus
      summary: Get airdrop on-chain status
      description: Returns the on-chain indexing status of an airdrop.
      tags:
        - Indexer
      parameters:
        - $ref: "#/components/parameters/AirdropIdParam"
      responses:
        "200":
          description: Airdrop on-chain status
          content:
            application/json:
              schema:
                type: object
                properties:
                  airdrop_id:
                    type: string
                  status:
                    type: string
                  events_count:
                    type: integer
        "400":
          description: Invalid airdrop ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Airdrop not indexed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/airdrops/{id}/onchain-recipients:
    get:
      operationId: getAirdropOnchainRecipients
      summary: Get on-chain recipients
      description: |
        Returns recipients derived from indexed on-chain claim events.
        Pagination is a slice of the fully-materialized list.
      tags:
        - Indexer
      parameters:
        - $ref: "#/components/parameters/AirdropIdParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Paginated on-chain recipients
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "400":
          description: Invalid airdrop ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/recipients/{address}/claims:
    get:
      operationId: getRecipientClaims
      summary: Get recipient claims
      description: Returns on-chain claim events for a specific recipient address.
      tags:
        - Indexer
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
          description: Stellar recipient address
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Paginated claim events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "400":
          description: Invalid recipient address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/indexer/status:
    get:
      operationId: getIndexerStatus
      summary: Indexer service status
      description: |
        Returns the current status, ledger height, and lag of the on-chain indexer.
        This endpoint is rate limited to 1 request per second per IP.
      tags:
        - Indexer
      responses:
        "200":
          description: Indexer status
          content:
            application/json:
              schema:
                type: object
                properties:
                  running:
                    type: boolean
                  latest_ledger:
                    type: integer
                    nullable: true
                  last_ledger:
                    type: integer
                    nullable: true
                  events_count:
                    type: integer
                  lag:
                    type: integer
                    nullable: true
                  ledger_lag:
                    type: integer
                    nullable: true
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
      description: |
        API key authentication. Provide your API key with the "Bearer " prefix in the
        `Authorization` header. Keys are created via `POST /api/v1/keys` and
        can be scoped to specific resources.

        Example: `Authorization: Bearer key_1234567890abcdef...`
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 1-indexed page number.
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Results per page (max 100).
    AirdropIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: ^drop_
      description: Airdrop ID
      example: drop_a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6
  schemas:
    StellarAddress:
      type: string
      pattern: ^G[A-Z0-9]{55}$
      description: Stellar public key (ed25519) starting with G
      example: GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335AX2OBFLDTQLNUEHRGPTM6RIA
...
    PaginationMeta:
      type: object
      description: |
        The canonical pagination envelope every list endpoint in this API
        returns (see the README's "Pagination" section) — `GET
        /api/v1/webhooks`, `/airdrops`, `/airdrops/:id/recipients`,
        `/alerts`, `/airdrops/:id/onchain-recipients`, and
        `/recipients/:address/claims`. `GET /api/v1/webhooks/:id/deliveries`
        is intentionally exempt (limit-only, no page/offset concept).
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        total_pages:
          type: integer
          example: 3
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
      required:
        - page
        - limit
        - total
        - total_pages
        - has_next
        - has_prev
...
    ErrorResponse:
      type: object
      description: >-
        Canonical error envelope. Clients should switch on `error.code`,
        which is stable, rather than on `error.message`, which may be
        reworded at any time.
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - PAYLOAD_TOO_LARGE
                - UNSUPPORTED_MEDIA_TYPE
                - RATE_LIMITED
                - INTERNAL_ERROR
                - UPSTREAM_ERROR
                - SERVICE_UNAVAILABLE
                - AIRDROP_NOT_FOUND
                - AIRDROP_NOT_INDEXED
                - RECIPIENT_LIMIT_EXCEEDED
                - CSV_INVALID_ENCODING
                - CSV_MISSING_COLUMNS
                - CSV_MALFORMED
                - CSV_EMPTY
                - WEBHOOK_NOT_FOUND
                - WEBHOOK_LIMIT_EXCEEDED
                - ALERT_NOT_FOUND
                - API_KEY_NOT_FOUND
                - PRICE_UNAVAILABLE
                - INDEXER_UNAVAILABLE
              example: WEBHOOK_NOT_FOUND
            message:
              type: string
              description: Human-readable error description
              example: Webhook not found
            request_id:
              type: string
              description: Correlates this response with server logs
              example: req_V1StGXR8Z5jdHi6BmyT
            details:
              type: object
              additionalProperties: true
              description: Optional structured context specific to the error code
          required:
            - code
            - message
      required:
        - error
    ApiKey:
      type: object
      properties:
        id:
          type: string
          pattern: ^key_
          example: key_1234567890abcdef
        label:
          type: string
          example: Production webhook worker
        key_prefix:
          type: string
          description: First eight characters of the generated API key for display only
          example: 3b28a419
        scopes:
          type: array
          items:
            type: string
          example:
            - webhooks:write
        tier:
          type: string
          enum:
            - free
            - pro
            - admin
          description: >-
            Rate limit tier for this key. Each key is metered in its own
            fixed-window bucket sized by its tier, so one key exhausting its
            limit does not affect any other key.
          example: free
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - label
        - key_prefix
        - scopes
        - created_at
        - last_used_at
    ApiKeyCreateRequest:
      type: object
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 120
          example: Production webhook worker
        scopes:
          type: array
          items:
            type: string
          default:
            - default
          example:
            - webhooks:write
        tier:
          type: string
          enum:
            - free
            - pro
            - admin
          default: free
          description: Rate limit tier for the new key. Defaults to free.
          example: pro
      required:
        - label
    ApiKeyCreateResponse:
      type: object
      properties:
        api_key:
          type: string
          description: Plaintext API key secret. Returned only once at creation time.
          example: 7a75e39a0fb34c0c90b2343f61b557aa7f0fd965baf0e5a97d30c384d83d1d75
        key:
          $ref: "#/components/schemas/ApiKey"
      required:
        - api_key
        - key
    ApiKeyListResponse:
      type: object
      properties:
        keys:
          type: array
          items:
            $ref: "#/components/schemas/ApiKey"
      required:
        - keys
    ApiKeyRevokeResponse:
      type: object
      properties:
        deleted:
          type: boolean
          example: true
        key:
          $ref: "#/components/schemas/ApiKey"
      required:
        - deleted
        - key
    ValidationErrorDetail:
      type: object
      properties:
        error:
          type: string
          example: Validation error
        message:
          type: string
          example: url must be a valid HTTP or HTTPS URL
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              issue:
                type: string
      required:
        - error
        - message
...
    AddRecipientsRequest:
      type: object
      properties:
        recipients:
          type: array
          description: Recipients to add to the airdrop
          items:
            $ref: "#/components/schemas/Recipient"
      required:
        - recipients
    Recipient:
      type: object
      properties:
        address:
          $ref: "#/components/schemas/StellarAddress"
        amount:
          type: number
          minimum: 0
          exclusiveMinimum: true
          description: Amount in asset units (at most 7 decimal places)
          example: 100.5
      required:
        - address
        - amount
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorDetail"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Missing or invalid API key
              request_id: req_V1StGXR8Z5jdHi6BmyT
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: WEBHOOK_NOT_FOUND
              message: Webhook endpoint not found
              request_id: req_V1StGXR8Z5jdHi6BmyT
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: RATE_LIMITED
              message: Rate limit exceeded. Try again in 42 seconds.
              request_id: req_V1StGXR8Z5jdHi6BmyT
              details:
                retry_after_seconds: 42
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Maximum requests per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when the window resets
        X-RateLimit-Tier:
          schema:
            type: string
          description: >-
            Rate limit tier the limit was applied from. Only present when the
            request was metered per API key rather than per IP.
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: An unexpected error occurred
              request_id: req_V1StGXR8Z5jdHi6BmyT